리뷰 상세정보 작업

+++ htashop/src/main/webapp/products/reviewDetail.jsp
@@ -17,22 +17,29 @@
 <%
 	// include 시킨 navbar의 nav-item 중에서 페이지에 해당하는 nav-item를 active 시키기위해서 "menu"라는 이름으로 페이지이름을 속성으로 저장한다.
 	// pageContext에 menu라는 이름으로 설정한 속성값은 navbar.jsp의 6번째 라인에서 조회해서 navbar의 메뉴들 중 하나를 active 시키기 위해서 읽어간다.
-	pageContext.setAttribute("menu", "review");
+//	pageContext.setAttribute("menu", "review");
 %>
 <%@include file="../common/navbar.jsp"%>
 <div class="container">
+
 	<div class="row justify-content-end">
 		<div  class="col-sm-2">
 			<%@ include file="../common/left.jsp" %>
 		</div>	
+		
 		<div class="col-sm-10 align-self-end mt-4">
 			<div class="row mb-3">
+				<div class="col">
+			<h1 class="fs-3">리뷰 상세</h1>
+		</div>	
 <%
 	// 클라이언트는 게시글 상세 페이지를 조회할 때 "detail.jsp?no=글번호&pageNo=페이지번호" 로 요청한다.(list.jsp의 78번 라인 참조)
 	// 요청파라미터에서 글번호와 페이지 번호를 조회한다.
 	int reviewNo = Integer.parseInt(request.getParameter("reviewNo"));
 	String pageNo = request.getParameter("pageNo");
 	String error = request.getParameter("error");
+	
+
 	
 	// 게시글 정보를 제공하는 BoardDao객체를 획득한다.
 	ReviewDao reviewDao = new ReviewDao();	
@@ -103,12 +110,24 @@
 			<div class="d-flex justify-content-between">
 				<div>
 <%
-//	if (loginedUserInfo != null &&  == reviewDto) {		// 로그인한 사용자의 사용자번호와 게시글작성자의 사용자번호가 일치하는 경우 버튼이 출력된다.
+	if (loginedUserInfo != null && loginedUserInfo.getUserNo()==reviewDto.getUserNo()) {		// 로그인한 사용자의 사용자번호와 게시글작성자의 사용자번호가 일치하는 경우 버튼이 출력된다.
 %>
-					<a href="delete.jsp?no=<%=reviewDto.getReviewNo() %>&pageNo=<%=pageNo %>" class="btn btn-danger">삭제</a>
+				<a href="delete.jsp?no=<%=reviewDto.getReviewNo()%>&pageNo=<%=pageNo %>" class="btn btn-danger">삭제</a>
+				<a href="modifyForm.jsp?no=<%=reviewDto.getReviewNo()%>" class="btn btn-danger">수정</a>					
 <%
-//	}
-%>				
+	}
+%>		
+<% 		
+	if (loginedUserInfo != null && "admin".equals(loginedUserInfo.getUserType())){
+	
+		%>	
+				<a href="replyForm.jsp?no=<%=reviewDto.getReviewNo() %>&pageNo=<%=pageNo %>" class="btn btn-primary">답글</a>
+				<a href="delete.jsp?no=<%=reviewDto.getReviewNo() %>&pageNo=<%=pageNo %>" class="btn btn-danger">삭제</a>	
+					
+<%
+	}
+%>
+
 				</div>
 				<a href="detail.jsp?no=<%=reviewDto.getProductNo() %>" class="btn btn-primary">목록</a>
 			</div>
